+++ examples/sglang_multiturn/config/tool_config/sandbox_fusion_tool_config.yaml
@@ -7,6 +7,7 @@ tools:
       rate_limit: 10
       default_timeout: 30
       default_language: "python"
+      type: native
 
     tool_schema:
       type: "function"
