+++ tests/workers/rollout/resource/tool_configs/mcp_tool_config
@@ -3,7 +3,7 @@ tools:
     config:
       rate_limit: 120
       timeout: 120
-      use_mcp: true
+      type: mcp
     mcp:
       mcp_servers_config_path: ./resource/tool_configs/mcp_server.json
       # optional
