+++ tests/workers/rollout/resource/tool_configs/sandbox_fusion_tool_config
@@ -2,6 +2,7 @@ tools:
   - class_name: "verl.tools.sandbox_fusion_tools.SandboxFusionTool"
     config: 
       sandbox_fusion_url: "https://xxx.apigateway-cn-beijing.volceapi.com/run_code"
+      type: native
     tool_schema:
       type: "function"
       function:
