+++ verl/workers/rollout/sglang_rollout/sglang_rollout.py
@@ -57,25 +57,13 @@ from transformers import PreTrainedTokenizer
 from verl import DataProto
 from verl.third_party.sglang import parallel_state as sglang_ps
 from verl.tools.base_tool import BaseTool
-from verl.tools.schemas import (
-    OpenAIFunctionCallSchema,
-    OpenAIFunctionParsedSchema,
-    OpenAIFunctionToolCall,
-)
-from verl.tools.utils.mcp_clients.McpClientManager import ClientManager
+from verl.tools.schemas import OpenAIFunctionCallSchema, OpenAIFunctionParsedSchema, OpenAIFunctionToolCall
+from verl.tools.utils.tool_registry import initialize_tools_from_config
 from verl.utils.debug import GPUMemoryLogger
 from verl.utils.net_utils import is_ipv6
-from verl.utils.torch_functional import (
-    get_response_mask,
-    pad_sequence_to_length,
-)
+from verl.utils.torch_functional import get_response_mask, pad_sequence_to_length
 from verl.workers.rollout.base import BaseRollout
-from verl.workers.rollout.schemas import (
-    AsyncRolloutRequest,
-    AsyncRolloutRequestStateEnum,
-    FinishReasonTypeEnum,
-    Message,
-)
+from verl.workers.rollout.schemas import AsyncRolloutRequest, AsyncRolloutRequestStateEnum, FinishReasonTypeEnum, Message
 from verl.workers.rollout.sglang_rollout.utils import broadcast_pyobj
 
 try:
@@ -427,71 +415,8 @@ class SGLangRollout(BaseRollout):
         if config.multi_turn.tool_config_path is None:
             return [], {}, None, [], None
 
-        import importlib.util
-        import sys
-
-        from omegaconf import OmegaConf
-
-        from verl.tools.schemas import OpenAIFunctionToolSchema
-
-        async def initialize_tools_from_config(tools_config: str):
-            tool_list = []
-            for tool_config in tools_config.tools:
-                cls_name = tool_config.class_name
-                module_name, class_name = cls_name.rsplit(".", 1)
-
-                if module_name not in sys.modules:
-                    spec = importlib.util.find_spec(module_name)
-                    module = importlib.util.module_from_spec(spec)
-                    sys.modules[module_name] = module
-                    spec.loader.exec_module(module)
-                else:
-                    module = sys.modules[module_name]
-
-                tool_cls = getattr(module, class_name)
-
-                use_mcp = tool_config.config.use_mcp if "use_mcp" in tool_config.config else False
-                if use_mcp:
-                    mcp_servers_config_path = tool_config.mcp.mcp_servers_config_path
-                    tool_selected_list = tool_config.mcp.tool_selected_list if "tool_selected_list" in tool_config.mcp else None
-                    await ClientManager.initialize(mcp_servers_config_path, tool_config.config.rate_limit)
-                    # Wait for MCP client to be ready
-                    max_retries = 10
-                    retry_interval = 2  # seconds
-                    for i in range(max_retries):
-                        tool_schemas = await ClientManager.fetch_tool_schemas(tool_selected_list)
-                        if tool_schemas:
-                            break
-                        if i < max_retries - 1:
-                            logger.debug(f"Waiting for MCP client to be ready, attempt {i + 1}/{max_retries}")
-                            await asyncio.sleep(retry_interval)
-                    else:
-                        raise RuntimeError("Failed to initialize MCP tools after maximum retries")
-                    # mcp registry
-                    assert len(tool_schemas), "mcp tool is empty"
-                    for tool_schema_dict in tool_schemas:
-                        logger.debug(f"tool_schema_dict: {tool_schema_dict}")
-                        tool_schema = OpenAIFunctionToolSchema.model_validate(tool_schema_dict)
-                        tool = tool_cls(
-                            config=OmegaConf.to_container(tool_config.config, resolve=True),
-                            tool_schema=tool_schema,
-                        )
-                        tool_list.append(tool)
-                else:
-                    tool_schema_dict = OmegaConf.to_container(tool_config.tool_schema, resolve=True)
-                    tool_schema = OpenAIFunctionToolSchema.model_validate(tool_schema_dict)
-                    tool = tool_cls(
-                        config=OmegaConf.to_container(tool_config.config, resolve=True),
-                        tool_schema=tool_schema,
-                    )
-                    tool_list.append(tool)
-
-            return tool_list
-
         tools_config_file = config.multi_turn.tool_config_path
-        tools_config = OmegaConf.load(tools_config_file)
-        loop = asyncio.get_event_loop()
-        tool_list = loop.run_until_complete(initialize_tools_from_config(tools_config))
+        tool_list = initialize_tools_from_config(tools_config_file)
 
         logger.info(f"Initialize tools from configuration.: tool_list: {tool_list}")
         tool_schemas = [tool.get_openai_tool_schema().model_dump() for tool in tool_list]
